Clean up dead code and brittle tests
