+++ launcher/tests/renderer-wiring.test.cjs
@@ -1,0 +1,76 @@
+const test = require("node:test");
+const assert = require("node:assert/strict");
+const fs = require("node:fs");
+const path = require("node:path");
+
+const launcherRoot = path.resolve(__dirname, "..");
+const appSource = fs.readFileSync(path.join(launcherRoot, "src", "App.tsx"), "utf8");
+const electronMain = fs.readFileSync(path.join(launcherRoot, "electron", "main.cjs"), "utf8");
+const browserHostSource = fs.readFileSync(path.join(launcherRoot, "electron", "browser-host.cjs"), "utf8");
+const preloadSource = fs.readFileSync(path.join(launcherRoot, "electron", "preload.cjs"), "utf8");
+
+test("embedded ChatGPT is measured only after its animated surface mounts", () => {
+  assert.match(appSource, /const \[browserSlot, setBrowserSlot\] = useState<HTMLDivElement \| null>\(null\)/);
+  assert.match(appSource, /setBrowserSurfaceActive\(browserSurfaceActive\)\.then\(\(\) => \{/);
+  assert.match(appSource, /observer\.observe\(browserSlot\)/);
+  assert.match(appSource, /ref=\{browserSlotRef\}/);
+});
+
+test("closing the launcher follows the persisted background-runtime preference", () => {
+  assert.match(
+    electronMain,
+    /if \(stateStore\.read\(\)\.keepRunningOnClose && tray\) window\.hide\(\);\s*else void requestQuit\(\);/,
+  );
+  assert.match(appSource, /setPreference\("keepRunningOnClose", checked\)/);
+});
+
+test("normal shutdown persists the ChatGPT session before closing browser views", () => {
+  const persist = electronMain.indexOf("await browserHost?.persistSession()");
+  const destroy = electronMain.indexOf("browserHost?.destroy()", persist);
+  assert.ok(persist >= 0, "shutdown must persist the ChatGPT session");
+  assert.ok(destroy > persist, "browser views must close only after session persistence completes");
+});
+
+test("the renderer bridge switch reaches the fail-closed runtime route", () => {
+  assert.match(appSource, /api!\.setBridgeEnabled\(enabled\)/);
+  assert.match(electronMain, /runtimeHost\.setBridgeEnabled\(enabled === true\)/);
+  assert.match(electronMain, /codexRestartRequired:\s*true/);
+});
+
+test("MCP connection remains unavailable until the model catalog is verified", () => {
+  assert.match(
+    appSource,
+    /snapshot\.state\.codexCatalogVerified \? copy\.mcpStepTwoHint : copy\.mcpCatalogRequired/,
+  );
+  assert.match(appSource, /\|\| !snapshot\.state\.codexCatalogVerified/);
+});
+
+test("MCP navigation remains locked while an operation is active", () => {
+  assert.match(appSource, /<McpSurface[\s\S]*?operation=\{operation\}/);
+  assert.match(appSource, /const busy = localBusy \|\| operation\?\.status === "running"/);
+  assert.match(appSource, /const safeMove = async \(next: number\) => \{\s*if \(busy\) return;/);
+  assert.match(appSource, /disabled=\{busy \|\| index > step\}/);
+});
+
+test("saved ChatGPT authentication is refreshed before setup is presented", () => {
+  assert.match(electronMain, /browserHost\.refreshAuthentication\(\)/);
+  assert.match(appSource, /browser\?\.status === "loading" \? copy\.checkingSignIn/);
+});
+
+test("completed model setup remains a repeatable capability probe", () => {
+  assert.match(appSource, /<SetupRow[\s\S]*?onAction=\{install\}[\s\S]*?repeatable/);
+  assert.match(appSource, /complete && !repeatable/);
+  assert.match(
+    electronMain,
+    /!setupState\.coreSetupComplete[\s\S]*?smokePassedThisSession[\s\S]*?smokePassedForCurrentVersion\(setupState\)/,
+  );
+});
+
+test("session reminders expose dismissal and a real storage-clearing logout", () => {
+  assert.match(electronMain, /sessionRefreshReminderAt:\s*nextSessionRefreshReminderAt\(\)/);
+  assert.match(electronMain, /launcher:session-reminder-dismiss/);
+  assert.match(electronMain, /launcher:browser-logout[\s\S]*?browserHost\.logout\(\)/);
+  assert.match(preloadSource, /dismissSessionReminder:[\s\S]*?launcher:session-reminder-dismiss/);
+  assert.match(preloadSource, /logoutChatGpt:[\s\S]*?launcher:browser-logout/);
+  assert.match(browserHostSource, /session\.clearStorageData\(\)/);
+});
