+++ src/tunnel.ts
@@ -372,7 +372,3 @@ export async function waitForTunnelReady(
   }
   return status;
 }
-
-export function tunnelClientVersion(): string {
-  return TUNNEL_VERSION;
-}
